Rewrite of components, new services added, utility changes
- New service authentication.py and limits.py to split components
- Moved user_required() and admin_required() to authentication.py
- Initialization of components from create_app() now in seperate files
- Optimization of imports
